requirements.txt
ffmpeg-python==0.2.0
future==1.0.0
pillow==11.2.1
pycryptodome==3.23.0
